do the actual split
